+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.1] / 2022-08-09
+### Features
+- Add `SetPanelsOrderByTitle` and `SetPanelsOrderBy`
+### Added
+- Add internal `ObservableCollectionExtension`
+### Updated
+- Update `MoveRibbonPanel`
+- Update Order methods
+- Update methods to `internal`
+
 ## [0.3.0] / 2022-08-08
 ### Features
 - Add `IExternalCommandAvailability` on `NewToggleButtonData` Command